Moves all fonts and assets into src
