--- lib/add-message/style.css
@@ -1,70 +1,0 @@
-@@font-face {
-  font-family: SourceSansLight;
-  src: url('../../fonts/SourceSansPro-Regular.otf');
-}
-
-.addMessageContainer {
-  bottom: 0;
-  position: fixed;
-  margin: 0;
-  margin-left: 0.2em;
-  right: 0;
-  left: 0;
-  width: 375px;
-}
-
-
-
-.arrow {
-  width: 3em;
-  height: 3em;
-  -webkit-filter: invert(1);
-   filter: invert(1);
-   padding-top: 0.2em;
-}
-
-.arrowContainer {
-  background-color: #F19024;
-  display: inline-block;
-  border-radius: 100%;
-  margin: 1em;
-  width: 3.2em;
-  height: 3.2em;
-  text-align: center;
-  position: absolute;
-  left: 17em;
-  top: 2em;
-}
-
-.message {
-  font-family: SourceSansLight;
-  background-color: #595959;
-  display: inline-block;
-  color: #ffffff;
-  padding: 1em;
-  padding-right: 2.5em;
-  padding-left: 1.8em;
-  padding-bottom: 1.3em;
-  border-radius: 8px;
-  margin-bottom: 1.2em;
-}
-
-@media (max-width: 321px) {
-  .addMessageContainer {
-    width: 320px;
-    padding: 0;
-  }
-  .message {
-    padding-left: 0.5em;
-    padding-right: 0.5em;
-  }
-  .arrowContainer {
-    left: 14.7em;
-  }
-}
-
-@media (min-width: 380px) {
-  .addMessageContainer {
-    display: none;
-  }
-}
