+++ lib/date-picker/index.js
@@ -1,5 +1,5 @@
 import React, { Component, PropTypes } from 'react';
-import DatePicker from '../datepicker/Datepicker';
+import DatePicker from '../react-date-picker/Datepicker';
 import moment from 'moment';
 import './styles.css';
 class Calendar extends Component {
