--- lib/datepicker/css/index.css
@@ -1,77 +1,0 @@
-.datepicker {
-  position: relative;
-}
-
-.calendar {
-  font-family: 'Trebuchet MS', sans-serif;
-  position: absolute;
-  top: 38px;
-  left: 3px;
-  width: 230px;
-  z-index: 9999;
-  border: solid 1px #eee;
-  background: #fff;
-}
-
-.calendar td {
-  text-align: center;
-}
-
-.calendar thead td {
-  padding: 10px 0;
-}
-
-.calendar__day {
-  font-size: 13px;
-  cursor: pointer;
-  padding: 7px 9px 6px;
-  border: solid 1px #eee;
-}
-
-.calendar__disabledDay {
-  opacity: 0.3;
-}
-
-.calendar__dayFromOtherMonth {
-  color: #ccc;
-}
-
-.calendar__dayInRange {
-  background: rgba(192, 232, 250, 0.5);
-}
-
-.calendar__activeDay {
-  background: rgba(192, 232, 250, 0.5);
-}
-
-.calendar__prevMonth,
-.calendar__nextMonth {
-  cursor: pointer;
-}
-
-.calendar__prevMonth:before {
-  content: '<';
-}
-
-.calendar__nextMonth:before {
-  content: '>';
-}
-
-.datepicker__input {
-  position: relative;
-  width: 100px;
-  margin: 5px;
-  padding: 5px 0;
-  border: 1px solid #d1d1d1;
-  display: inline-block;
-  font-weight: 400;
-  text-align: center;
-  vertical-align: middle;
-  white-space: nowrap;
-  border-radius: 4px;
-  background-color: #f0f0f0;
-  cursor: pointer;
-  color: transparent;
-  text-shadow: 0 0 0 #666;
-  outline: none;
-}
