+++ lib/filter-tile/description.js
