+++ lib/filter-tile/index.js
