+++ lib/filter-tile/infoText.js
