+++ lib/hero-image-header/index.js
@@ -1,7 +1,7 @@
 import React, { Component, PropTypes } from 'react';
 require('./style.css');
-const signpost = require('./assets/goingplaces.svg');
-import SearchBar from '../search';
+const signpost = require('../../src/assets/goingplaces.svg');
+import SearchBar from '../search-bar';
 export default class Header extends Component {
   render () {
     const {
