+++ lib/hero-image-header/style.css
