+++ lib/hero-image-header/test/index.test.js
