+++ lib/react-date-picker/DateInput.js
@@ -5,12 +5,14 @@ export default class DateInput extends Component {
   render () {
     const { inputValue, inputOnClick, inputClassName } = this.props;
 
-    return <input type='text'
-                  ref='dateInput'
-                  className={inputClassName}
-                  value={inputValue.format('YYYY-MM-DD')}
-                  onClick={inputOnClick}
-                  readOnly />;
+    return <input
+              type='text'
+              ref='dateInput'
+              className={inputClassName}
+              value={moment(inputValue).format('YYYY-MM-DD')}
+              onClick={inputOnClick}
+              readOnly
+            />;
   }
 }
