+++ lib/react-date-picker/Datepicker.js
@@ -1,40 +1,15 @@
 import React, { Component, PropTypes } from 'react';
-import ReactDOM from 'react-dom';
 import moment from 'moment';
 import DateInput from './DateInput';
 import Calendar from './Calendar';
 
 export default class DatePicker extends Component {
   constructor (props) {
-    super(props);
-
+    super(props); // http://stackoverflow.com/q/30571875/1148249
     this.state = {
       date: moment(props.date),
       isCalendarOpen: false
     };
-  }
-
-  componentWillMount () {
-    document.addEventListener('mousedown', this.onClickOutside.bind(this));
-    document.addEventListener('touchstart', this.onClickOutside.bind(this));
-  }
-
-  componentWillUnmount () {
-    document.removeEventListener('mousedown', this.onClickOutside.bind(this));
-    document.removeEventListener('touchstart', this.onClickOutside.bind(this));
-  }
-
-  onClickOutside (e) {
-    const calendar = ReactDOM.findDOMNode(this.refs.calendar);
-    const dateInput = ReactDOM.findDOMNode(this.refs.dateInput);
-
-    if (!calendar) return;
-
-    if (!calendar.contains(e.target) && !dateInput.contains(e.target)) {
-      this.setState({
-        isCalendarOpen: false
-      });
-    }
   }
 
   toggleCalendar () {
@@ -62,7 +37,7 @@ export default class DatePicker extends Component {
     }
 
     return <Calendar ref='calendar'
-                     date={moment(this.props.date)}
+                     date={this.state.date}
                      minDate={minDate}
                      maxDate={maxDate}
                      selectDay={this.selectDay.bind(this)}
@@ -80,7 +55,8 @@ export default class DatePicker extends Component {
     const { clickOnDate, name } = this.props;
     if (clickOnDate) {
       this.setState({
-        isCalendarOpen: false
+        isCalendarOpen: false,
+        date: date
       });
       return clickOnDate(date, name);
     }
@@ -96,11 +72,12 @@ export default class DatePicker extends Component {
 
     return (
       <div className={datepickerClassName}>
-        <DateInput ref='dateInput'
-                   inputValue={this.state.date}
-                   inputOnClick={this.toggleCalendar.bind(this)}
-                   inputClassName={inputClassName} />
-
+        <DateInput
+          ref='dateInput'
+          inputValue={this.state.date}
+          inputOnClick={this.toggleCalendar.bind(this)}
+          inputClassName={inputClassName}
+        />
         {this.renderCalendar()}
       </div>
     );
