+++ lib/react-date-picker/Day.js
@@ -20,7 +20,7 @@ const Day = ({
 
   let classes = [dayClassName];
 
-  if (day.isSame(active, 'day')) {
+  if (moment(day).isSame(active, 'day')) {
     classes.push(dayActiveClassName);
   }
 
@@ -32,7 +32,7 @@ const Day = ({
     classes.push(dayFromOtherMonthClassName);
   }
 
-  return <td className={classes.join(' ')} onClick={handleOnClick}>{ day.date() }</td>;
+  return <td className={classes.join(' ')} onClick={handleOnClick}>{ moment(day).date() }</td>;
 };
 
 Day.propTypes = {
