+++ lib/react-date-picker/test/calendar.test.js
