+++ lib/react-date-picker/test/dateInput.test.js
