+++ lib/react-date-picker/test/day.test.js
