+++ lib/react-date-picker/test/index.test.js
