+++ lib/react-date-picker/utils/index.js
@@ -21,6 +21,20 @@ export function weekEnum (daysOfMonth) {
   return results;
 }
 
+export function isDateInRange (currentDate, date, startDate, endDate, disabled) {
+  let isInRange = date.isSame(currentDate);
+
+  if (startDate) {
+    isInRange = date.isSame(currentDate) || date.isAfter(currentDate) || date.isSame(startDate);
+  }
+
+  if (endDate) {
+    isInRange = date.isSame(currentDate) || date.isBefore(currentDate) || date.isSame(endDate);
+  }
+
+  return !disabled ? isInRange : false;
+}
+
 export function isDateFromNextMonth (date, displayedMonth) {
   return date.year() > displayedMonth.year() || date.year() ===
          displayedMonth.year() && date.month() > displayedMonth.month();
