+++ lib/search-bar/index.js
