+++ lib/search-bar/style.css
