+++ lib/search-summary/edit-details/index.js
@@ -1,7 +1,7 @@
 import React, { Component, PropTypes } from 'react';
-import DropDown from '../../drop-down';
+import DropDown from '../../select-drop-down';
 import Calendar from '../../date-picker';
-import '../../datepicker/css/index.css';
+import '../../react-date-picker/css/index.css';
 import {
   adultOptions,
   childOptions,
@@ -9,7 +9,7 @@ import {
   durationOptions,
   childAgeOptions
 } from './options.js';
-const closeImage = require('../assets/close002.svg');
+const closeImage = require('../../../src/assets/close.svg');
 import './style.css';
 export default class EditDetails extends Component {
   render () {
