+++ lib/search-summary/header/index.js
@@ -1,6 +1,6 @@
 import React, { Component, PropTypes } from 'react';
 import './style.css';
-const postImage = require('../assets/goingplaces.svg');
+const postImage = require('../../../src/assets/goingplaces.svg');
 export default class SearchSummaryHeader extends Component {
   render () {
     const { durationInWeeks, paxMix, changeDetails, viewTags } = this.props;
