+++ lib/search-summary/index.js
@@ -3,7 +3,7 @@ import SearchSummaryHeader from './header';
 import EditDetails from './edit-details';
 import TagView from './tag-view';
 import './style.css';
-import '../datepicker/css/index.css';
+import '../react-date-picker/css/index.css';
 
 export default class SearchSummary extends Component {
   constructor () {
