+++ lib/search-summary/tag-view/index.js
@@ -1,5 +1,5 @@
 import React, { Component, PropTypes } from 'react';
-const closeImage = require('../assets/close002.svg');
+const closeImage = require('../../../src/assets/close.svg');
 
 export default class TagView extends Component {
   render () {
